pyproject.toml
[project]
name = "oma"
version = "0.1.0"
description = "Oracle Migration Assistant - AI-powered Oracle to PostgreSQL SQL converter"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=1.24.0",
    "boto3>=1.42.0",
    "defusedxml>=0.7.1",
    "sqlalchemy>=2.0.0",
    "rich>=13.0.0",
]
